Time: 0 ms (100%), Space: 8.7 MB (26.66%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if (x<0) return false ;
-       int  n= x;
+   long    int  n= x;
      long  int  re= 0;
         while (x>0){
             int r= x%10;
